+++ tests/internals/test_azure_blob_storage.py
@@ -4,7 +4,12 @@ from azure.storage.blob import BlobType
 
 from cshelve._factory import factory
 from azure.core.exceptions import ResourceNotFoundError
-from cshelve import AuthArgumentError, AuthTypeError, KeyNotFoundError
+from cshelve import (
+    AuthArgumentError,
+    AuthTypeError,
+    KeyNotFoundError,
+    ConfigurationError,
+)
 
 
 @patch("azure.identity.DefaultAzureCredential")
@@ -115,6 +120,24 @@ def test_missing_env_var(auth_type):
     "auth_type",
     ["access_key", "connection_string"],
 )
+def test_missing_container(auth_type):
+    """
+    Ensure a ConfigurationError is raised when the container name is missing from the configuration.
+    """
+    config = {
+        "account_url": "https://account.blob.core.windows.net",
+        "auth_type": auth_type,
+    }
+
+    with pytest.raises(ConfigurationError):
+        provider = factory("azure-blob")
+        provider.configure(config)
+
+
+@pytest.mark.parametrize(
+    "auth_type",
+    ["access_key", "connection_string"],
+)
 def test_missing_env_var_value(auth_type):
     """
     Test the Azure Blob Storage client with the connection string/account key authentication without providing the
@@ -139,6 +162,7 @@ def test_wrong_auth_type():
     config = {
         "account_url": "https://account.blob.core.windows.net",
         "auth_type": "unknonwn",
+        "container_name": "container",
     }
 
     with pytest.raises(AuthTypeError):